creating get by id logic for scenaries and some other small changes
