+++ src/app/admin/create-scenario/[id]/page.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { use } from 'react';
+import { CreateScenarioBase } from '../_components/CreateScenarioBase';
+
+export default function Page({ params }: { params: Promise<{ id: string }> }) {
+  const { id } = use(params);
+  return <CreateScenarioBase scenarioId={id} />;
+}
